pyproject.toml

[project]
name = "jobsearchlangchain"
version = "0.1.0"
dependencies = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.2.1",
]

[tool.uv]
requires-python = ">=3.9"
